Find missing input pieces.

find_tests_without_reqs() returns IDs from the test results that are not associated with a requirement. find_reqs_with_missing_tests() goes in the reverse direction, returning requirements with test IDs that aren't found in the test results. find_reqs_without_stories() returns requirements that are not linked to a story.

find_missing() is a convenience wrapper around all of the above functions. It returns their results in a list and also displays messages about the results.

find_missing(merged_inputs)

find_tests_without_reqs(merged_inputs)

find_reqs_with_missing_tests(merged_inputs)

find_reqs_without_stories(merged_inputs)

Arguments

merged_inputs
Tibble with stories, requirements and tests, as returned by create_validation_docs().

Value

Tibble with the missing items, or, in the case of find_missing(), a list of tibbles.

Details

The requirement columns (RequirementId and RequirementDescription) are optional (see input_formats). If not present, find_tests_without_reqs() and find_reqs_with_missing_tests() use stories in place of requirements. find_reqs_without_stories(), on the other hand, errors if the columns aren't present because it can't do anything sensible without them.

See also